[#200] contextmenu 피드백 반영 > 하나의 재귀에서 둘로 명시적 분리 ################################ - 피드백 반영
